Move app.py back to root directory; rename package from app to server to prevent identical names; relegate __init__.py to initializing the Flask app only
